// apps/oberon360-ic/src/types/CustomerResponse.ts
export type CustomerResponse = {
  cliente_ID: string;
  cliente_Nombre: string;
  servicioS_Servicio_Descripcion: string | null;
  servicioS_Servicio_ID: string | null;
  paqueteS_Paquete_ID: string | null;
  paqueteS_Paquete_Nombre: string | null;
  paqueteS_Servicio_Descripcion: string | null;
  paqueteS_Servicio_ID: string | null;
};